chore: bump to AGP 9.1.1 / Gradle 9.4.1 / Kotlin 2.3.20 / Hilt 2.59.2
AGP 9 ships Kotlin built-in; removed org.jetbrains.kotlin.android from
the plugin alias list. KSP 2.3.6 (KSP is now decoupled from the Kotlin
version string). Daemon heap bumped to 6 GB.

Kotlin 2.3 surfaced real issues; fixed at the source (no @Suppress):
- dead elvis (`?: return`/`?: continue`) on non-null CarStatus
- redundant .toFloat() on math.floor/ceil results already Float
- redundant `as ApiResult.Error` after smart-cast in BatteryViewModel
- unnecessary `?.` on smart-cast non-null status.state in DashboardScreen
- DisposableEffect block in two Fullscreen chart dialogs whose
  `Context as? Dialog` cast always failed — dead code removed, runtime
  behavior preserved (driven by DialogProperties already).

Known remaining: one Gradle-9 deprecation `ReportingExtension.file(String)`
emitted by detekt 1.23.8 at config time; detekt 2.0-alpha.2 fixes it but
requires `android.newDsl=false` which conflicts with AGP 9 defaults —
deferred until detekt 2.0 stable.

+++ app/src/main/java/com/matedroid/ui/screens/charges/ChargeDetailScreen.kt
@@ -823,8 +823,8 @@ private fun TemperatureChartCard(
 ) {
     val temps = chargePoints.mapNotNull { it.outsideTemp?.toFloat() }
     if (temps.size < 2) return
-    var yMin = (kotlin.math.floor(temps.min() ) ).toFloat()
-    var yMax = (kotlin.math.ceil(temps.max() ) ).toFloat()
+    var yMin = kotlin.math.floor(temps.min())
+    var yMax = kotlin.math.ceil(temps.max())
     if (yMin == yMax) {
         yMin -= 1
         yMax += 1
